# docs/api/minic-web-openapi.yaml
openapi: 3.0.3
info:
  title: MiniC Web Workbench API
  version: 0.1.0
  description: Browser/server API for the MiniC Visual Workbench.
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/compile/sessions:
    post:
      operationId: createCompileSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Compile session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionCreatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/compile/sessions/{sessionId}:
    delete:
      operationId: closeCompileSession
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Compile session closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionClosedResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/debug/sessions:
    post:
      operationId: createDebugSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Debug session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionCreatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/debug-sessions:
    post:
      operationId: createDebugSessionSpecAlias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Debug session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionCreatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/debug/sessions/{sessionId}:
    delete:
      operationId: closeDebugSession
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug session closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionClosedResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug-sessions/{sessionId}:
    delete:
      operationId: closeDebugSessionSpecAlias
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug session closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionClosedResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug/sessions/{sessionId}/source:
    post:
      operationId: updateDebugSource
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "200":
          description: Debug source updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebugSnapshotResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug-sessions/{sessionId}/source:
    post:
      operationId: updateDebugSourceSpecAlias
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "200":
          description: Debug source updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebugSnapshotResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug/sessions/{sessionId}/start:
    post:
      operationId: startDebugSession
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebugSnapshotResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/debug-sessions/{sessionId}/start:
    post:
      operationId: startDebugSessionSpecAlias
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebugSnapshotResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/debug/sessions/{sessionId}/breakpoints:
    post:
      operationId: addDebugBreakpoint
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BreakpointRequest"
      responses:
        "200":
          description: Debug breakpoint added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugStateDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug-sessions/{sessionId}/breakpoints/{line}:
    post:
      operationId: addDebugBreakpointSpecAlias
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/Line"
      responses:
        "200":
          description: Debug breakpoint added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugStateDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: removeDebugBreakpointSpecAlias
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/Line"
      responses:
        "200":
          description: Debug breakpoint removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugStateDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug/sessions/{sessionId}/breakpoints/{line}:
    delete:
      operationId: removeDebugBreakpoint
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/Line"
      responses:
        "200":
          description: Debug breakpoint removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugStateDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug/sessions/{sessionId}/commands/{command}:
    post:
      operationId: runDebugCommand
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/Command"
      responses:
        "200":
          description: Debug command result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugStateDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/debug-sessions/{sessionId}/{command}:
    post:
      operationId: runDebugCommandSpecAlias
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/Command"
      responses:
        "200":
          description: Debug command result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugStateDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/debug/sessions/{sessionId}/state:
    get:
      operationId: getDebugState
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugStateDto"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug-sessions/{sessionId}/state:
    get:
      operationId: getDebugStateSpecAlias
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugStateDto"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug/sessions/{sessionId}/views/metadata:
    get:
      operationId: getDebugMetadataView
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug metadata view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugMetadataViewDto"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug-sessions/{sessionId}/views/metadata:
    get:
      operationId: getDebugMetadataViewSpecAlias
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug metadata view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugMetadataViewDto"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug/sessions/{sessionId}/views/ast:
    get:
      operationId: getDebugAstView
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug AST view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugAstViewDto"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug-sessions/{sessionId}/views/ast:
    get:
      operationId: getDebugAstViewSpecAlias
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug AST view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugAstViewDto"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug/sessions/{sessionId}/views/ir:
    get:
      operationId: getDebugIrView
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug IR view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugIrViewDto"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug-sessions/{sessionId}/views/ir:
    get:
      operationId: getDebugIrViewSpecAlias
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug IR view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugIrViewDto"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug/sessions/{sessionId}/views/asm:
    get:
      operationId: getDebugAsmView
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug assembly view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugAsmViewDto"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug-sessions/{sessionId}/views/asm:
    get:
      operationId: getDebugAsmViewSpecAlias
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug assembly view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugAsmViewDto"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug/sessions/{sessionId}/views/data-structure:
    get:
      operationId: getDebugDataStructureView
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug data-structure view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugDataStructureViewDto"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug-sessions/{sessionId}/views/data-structure:
    get:
      operationId: getDebugDataStructureViewSpecAlias
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Debug data-structure view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiDebugDataStructureViewDto"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug/sessions/{sessionId}/snapshot:
    get:
      operationId: getDebugSnapshot
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Aggregate debug snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebugSnapshotResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/debug-sessions/{sessionId}/snapshot:
    get:
      operationId: getDebugSnapshotSpecAlias
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Aggregate debug snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebugSnapshotResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
    Command:
      name: command
      in: path
      required: true
      schema:
        type: string
    Line:
      name: line
      in: path
      required: true
      schema:
        type: integer
        format: int32
        minimum: 1
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/WebError"
    NotFound:
      description: Session or resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/WebError"
    Conflict:
      description: Command conflicts with current session state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/WebError"
  schemas:
...
    WebError:
      type: object
      required: [code, message, status]
      properties:
        code:
          type: string
        message:
          type: string
        status:
          type: integer
          format: int32
    CreateSessionRequest:
      type: object
      required: [sourceName, sourceText]
      properties:
        sourceName:
          type: string
        sourceText:
          type: string
    SessionCreatedResponse:
      type: object
      required: [sessionId, version]
      properties:
        sessionId:
          type: string
        version:
          type: integer
          format: int64
    SessionClosedResponse:
      type: object
      required: [sessionId, version, closed]
      properties:
        sessionId:
          type: string
        version:
          type: integer
          format: int64
        closed:
          type: boolean
...
    BreakpointRequest:
      type: object
      required: [line]
      properties:
        line:
          type: integer
          format: int32
          minimum: 1
...
    DebugSnapshotResponse:
      type: object
      required: [state, metadata, ast, ir, asm, dataStructure]
      properties:
        state:
          $ref: "#/components/schemas/UiDebugStateDto"
        metadata:
          $ref: "#/components/schemas/UiDebugMetadataViewDto"
        ast:
          $ref: "#/components/schemas/UiDebugAstViewDto"
        ir:
          $ref: "#/components/schemas/UiDebugIrViewDto"
        asm:
          $ref: "#/components/schemas/UiDebugAsmViewDto"
        dataStructure:
          $ref: "#/components/schemas/UiDebugDataStructureViewDto"
...
    UiSourceSpanDto:
      type: object
      required: [sourceName, startLine, startColumn, endLine, endColumn, startOffset, endOffset]
      properties:
        sourceName:
          type: string
        startLine:
          type: integer
          format: int32
        startColumn:
          type: integer
          format: int32
        endLine:
          type: integer
          format: int32
        endColumn:
          type: integer
          format: int32
        startOffset:
          type: integer
          format: int32
        endOffset:
          type: integer
          format: int32
...
    UiAstNodeVisualDto:
      type: object
      required: [id, kind, label, children, active]
      properties:
        id:
          type: string
        kind:
          type: string
        label:
          type: string
        range:
          nullable: true
          $ref: "#/components/schemas/UiSourceSpanDto"
        children:
          type: array
          items:
            $ref: "#/components/schemas/UiAstNodeVisualDto"
        active:
          type: boolean
...
    UiIrLineVisualDto:
      type: object
      required: [lineNumber, text, active]
      properties:
        lineNumber:
          type: integer
          format: int32
        text:
          type: string
        range:
          nullable: true
          $ref: "#/components/schemas/UiSourceSpanDto"
        active:
          type: boolean
...
    UiDebugStateDto:
      type: object
      required: [sourceName, executionState, currentSnapshot, snapshots, events, breakpoints]
      properties:
        sourceName:
          type: string
        executionState:
          type: string
        currentSnapshot:
          $ref: "#/components/schemas/UiDebugSnapshotDto"
        snapshots:
          type: array
          items:
            $ref: "#/components/schemas/UiDebugSnapshotDto"
        events:
          type: array
          items:
            $ref: "#/components/schemas/UiDebugEventDto"
        breakpoints:
          type: array
          items:
            $ref: "#/components/schemas/UiDebugBreakpointDto"
    UiDebugSnapshotDto:
      type: object
      required: [snapshotId, visibleStepIndex, functionName, blockLabel, instructionId, callStackSummary, processSpace, breakpointHit, stopReason]
      properties:
        snapshotId:
          type: integer
          format: int64
        visibleStepIndex:
          type: integer
          format: int64
        functionName:
          type: string
        blockLabel:
          type: string
        instructionId:
          type: string
        sourceRange:
          nullable: true
          $ref: "#/components/schemas/UiSourceSpanDto"
        callStackSummary:
          type: array
          items:
            type: string
        processSpace:
          $ref: "#/components/schemas/UiDebugProcessSpaceDto"
        breakpointHit:
          type: boolean
        stopReason:
          type: string
    UiDebugBreakpointDto:
      type: object
      required: [line, enabled]
      properties:
        line:
          type: integer
          format: int32
        enabled:
          type: boolean
    UiDebugEventDto:
      type: object
      required: [eventId, snapshotId, type, title, description, affectedValueRefs]
      properties:
        eventId:
          type: integer
          format: int64
        snapshotId:
          type: integer
          format: int64
        type:
          type: string
        title:
          type: string
        description:
          type: string
        sourceRange:
          nullable: true
          $ref: "#/components/schemas/UiSourceSpanDto"
        affectedValueRefs:
          type: array
          items:
            type: string
    UiDebugFrameDto:
      type: object
      required: [frameId, functionName, parameters, locals]
      properties:
        frameId:
          type: string
        functionName:
          type: string
        parameters:
          type: array
          items:
            $ref: "#/components/schemas/UiDebugVariableDto"
        locals:
          type: array
          items:
            $ref: "#/components/schemas/UiDebugVariableDto"
        returnTarget:
          type: string
        activeRange:
          nullable: true
          $ref: "#/components/schemas/UiSourceSpanDto"
    UiDebugVariableDto:
      type: object
      required: [name, address, typeName, valueKind, valueSummary, pointerTarget, typeShape, highlightedChange, explanation, fields, elements]
      properties:
        name:
          type: string
        address:
          type: string
        typeName:
          type: string
        valueKind:
          type: string
        valueSummary:
          type: string
        pointerTarget:
          type: string
        typeShape:
          type: string
        highlightedChange:
          type: boolean
        explanation:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/UiDebugVariableDto"
        elements:
          type: array
          items:
            $ref: "#/components/schemas/UiDebugVariableDto"
    UiDebugMetadataViewDto:
      type: object
      required: [executionState, stopReason, currentFunction, callStack, variables, stdout, stderr, breakpoints, events, timeline]
      properties:
        executionState:
          type: string
        stopReason:
          type: string
        currentFunction:
          type: string
        currentSourceRange:
          nullable: true
          $ref: "#/components/schemas/UiSourceSpanDto"
        callStack:
          type: array
          items:
            $ref: "#/components/schemas/UiDebugFrameDto"
        variables:
          type: array
          items:
            $ref: "#/components/schemas/UiDebugVariableDto"
        stdout:
          type: string
        stderr:
          type: string
        breakpoints:
          type: array
          items:
            $ref: "#/components/schemas/UiDebugBreakpointDto"
        events:
          type: array
          items:
            $ref: "#/components/schemas/UiDebugEventDto"
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/UiDebugTimelineItemDto"
    UiDebugAstViewDto:
      type: object
      required: [root, relatedIrIds, relatedAsmIds]
      properties:
        root:
          $ref: "#/components/schemas/UiAstNodeVisualDto"
        activeNode:
          nullable: true
          $ref: "#/components/schemas/UiDebugAstNodeDetailDto"
        relatedIrIds:
          type: array
          items:
            type: string
        relatedAsmIds:
          type: array
          items:
            type: string
    UiDebugAstNodeDetailDto:
      type: object
      required: [nodeId, kind, label, explanation]
      properties:
        nodeId:
          type: string
        kind:
          type: string
        label:
          type: string
        sourceRange:
          nullable: true
          $ref: "#/components/schemas/UiSourceSpanDto"
        explanation:
          type: string
    UiDebugIrViewDto:
      type: object
      required: [lines, currentInstructionId, explanation, operands]
      properties:
        lines:
          type: array
          items:
            $ref: "#/components/schemas/UiIrLineVisualDto"
        currentInstructionId:
          type: string
        currentSourceRange:
          nullable: true
          $ref: "#/components/schemas/UiSourceSpanDto"
        explanation:
          type: string
        operands:
          type: array
          items:
            $ref: "#/components/schemas/UiDebugIrOperandDto"
    UiDebugIrOperandDto:
      type: object
      required: [name, typeName, valueSummary, valueRef]
      properties:
        name:
          type: string
        typeName:
          type: string
        valueSummary:
          type: string
        valueRef:
          type: string
    UiDebugAsmViewDto:
      type: object
      required: [lines, explanation, relatedIrIds]
      properties:
        lines:
          type: array
          items:
            $ref: "#/components/schemas/UiAssemblyLineVisualDto"
        explanation:
          type: string
        relatedIrIds:
          type: array
          items:
            type: string
    UiDebugDataStructureViewDto:
      type: object
      required: [processSpace, visuals, warnings]
      properties:
        processSpace:
          $ref: "#/components/schemas/UiDebugProcessSpaceDto"
        visuals:
          type: array
          items:
            $ref: "#/components/schemas/UiDebugVisualStructureDto"
        warnings:
          type: array
          items:
            type: string
...
    UiDebugTimelineItemDto:
      type: object
      required: [snapshotId, visibleStepIndex, stopReason, breakpointHit]
      properties:
        snapshotId:
          type: integer
          format: int64
        visibleStepIndex:
          type: integer
          format: int64
        stopReason:
          type: string
        breakpointHit:
          type: boolean
        sourceRange:
          nullable: true
          $ref: "#/components/schemas/UiSourceSpanDto"